+++ android/app/build.gradle
@@ -54,7 +54,7 @@ android {
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
         minSdkVersion flutter.minSdkVersion
         targetSdkVersion flutter.targetSdkVersion
-        versionCode 10
+        versionCode 14
         versionName flutterVersionName
         multiDexEnabled true
         minSdkVersion 21
