+++ pubspec.yaml
@@ -16,7 +16,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.2
+build: 1
 
 environment:
   sdk: '>=3.2.3 <4.0.0'
@@ -54,7 +55,7 @@ dependencies:
   readmore: ^2.2.0
   skeletonizer: ^1.0.1
   cached_network_image: ^3.3.1
-  intl: ^0.18.1
+  # intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
